+++ .storybook/main.ts
@@ -12,11 +12,25 @@ const config: StorybookConfig = {
   ],
   framework: {
     name: '@storybook/react-webpack5',
-    options: {},
+    options: {
+      builder: {
+        useSWC: true,
+      },
+    },
   },
+  swc: () => ({
+    jsc: {
+      transform: {
+        react: {
+          runtime: 'automatic',
+        },
+      },
+    },
+  }),
   docs: {
     autodocs: true,
   },
+
   webpackFinal: async (config) => {
     config.resolve ??= {};
     config.resolve.plugins ??= [];