+++ src/components/Typography/index.ts
@@ -1,2 +1,3 @@
 // export * from './text';
 export * from './typography';
+export * from './defaultTypographyMixin';