+++ src/components/input/EditMode/stories/EditMode.stories.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 import type { Meta, StoryFn } from '@storybook/react';
 import { useGlobals } from '@storybook/preview-api';
@@ -95,9 +94,6 @@ export default {
       control: { type: 'boolean' },
     },
     skeleton: {
-      control: { type: 'boolean' },
-    },
-    isLoading: {
       control: { type: 'boolean' },
     },
     multilineView: {