+++ src/components/input/SliderInput/stories/SliderInput.stories.tsx
@@ -102,9 +102,6 @@ export default {
     showTooltip: {
       control: { type: 'boolean' },
     },
-    isLoading: {
-      control: { type: 'boolean' },
-    },
     icons: {
       control: false,
     },